Add some more parser test cases
The tests are mostly inspired by If8e96d9, and need to be updated when
that patch is merged. This is intentional. As the tests are now, they
are testing the current behavior. The patch If8e96d9 is *changing* the
current behavior. I find it important to make this visible, e.g. via
tests.

This patch also does:
* Make use of the …::class feature where possible.
* Remove obsolete `return true` from hook handler.
* Remove misplaced double escape from regular expression. Luckily this
  doesn't do anything, it's just a little confusing to read.
* Rewrite overly complicated condition.

Change-Id: I79cdb77e0469a62a83c9d97d8880376d1370ea9d