Allow overriding the link text
Displaying the character to be inserted is perfect in most cases, but
some obscure characters may require explanation or very long snippets
like templates a shorter label.

This patch allows setting a different label using the `label` XML-style
attribute. This attribute is technically usable with charinsert tags
containing multiple inserts, but it shouldn’t be used in that way: it
results in multiple links with the exact same text, so the user has no
way to distinguish them before clicking. (I couldn’t find any obvious
way to sytnactically prevent this, if anyone has an idea, I’m open to
it.)

To easily pass the attribute to where it’s used, the CharInsert class
is now instantiable, and the attributes are passed in the constructor.
While I was there, I also changed these methods to private and renamed
them to not have fully-qualified names, as they’re qualified by the
class name anyway.

There are two new parser test cases: one for the new functionality, and
one for the already-existing functionality involving the `+` character.
The latter describes how it actually works, which may not be how it was
planned to work, but I didn’t want to change the behavior in this patch.

Bug: T25354
Change-Id: I70d282e61d92fe8d82afa3d2443add6e36956dc1